#13 Refactorings to streamline namings and handling of getting and setting of config values
