+++ bin/init.js
@@ -1,5 +1,5 @@
 import powershell from './common/helper/powershell.js';
-import projectConfigFacade from "./configuration/facades/projectConfigFacade.js";
+import projectConfigFacade from "./configuration/facades/project-config-facade.js";
 import versionChecker from './common/helper/version-checker.js';
 import configValidator from './common/helper/config-validator.js';
